+++ Sources/TesseraTerminalSnapshotSupport/VirtualTerminalError.swift
@@ -1,9 +1,9 @@
-#if !os(Windows)
+#if canImport(CGhosttyVT)
   import CGhosttyVT
 #endif
 
 enum VirtualTerminalError: Error {
-  #if !os(Windows)
+  #if canImport(CGhosttyVT)
     case ghostty(operation: String, result: GhosttyResult)
   #endif
   case invalidSize(cols: Int, rows: Int)